% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemma_tools.R
\name{calc_kinship}
\alias{calc_kinship}
\title{Call gemma to calculate kinship matrix}
\usage{
calc_kinship(genotypes, annot, exec, chrname, basedir, phenofile)
}
\arguments{
\item{genotypes}{A data.table with genotypes of all mice}

\item{annot}{Annotations of markers, including chr column}

\item{exec}{gemma executable}

\item{chrname}{chr for LOCO}

\item{basedir}{base directory to write files to. Will override existing files}

\item{phenofile}{a file with phenotypes matching the genotypes table}
}
\value{
The kinship file name
}
\description{
Call gemma to calculate kinship matrix
}
