% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pylmm.R
\name{calc_pylmm_kinship}
\alias{calc_pylmm_kinship}
\title{Calculate kinship using PyLMM}
\usage{
calc_pylmm_kinship(genotypes, annot, pylmm_kinship, chrname, basedir)
}
\arguments{
\item{basedir}{}
}
\value{

}
\description{
Calculate kinship using PyLMM
}
