% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemma_tools.R
\name{combine_metaSOFT}
\alias{combine_metaSOFT}
\title{Combine phenotypes using metasoft}
\usage{
combine_metaSOFT(
  basedir,
  infiles,
  midfile,
  outfile,
  runit = FALSE,
  version = "2.0.1",
  xargs = ""
)
}
\arguments{
\item{infiles}{a vector of input files, each for one phenotype}

\item{outfile}{Output file name}

\item{runit}{Actually run metasoft. Otherwise only paste the files}

\item{version}{Metasoft version to download}

\item{xargs}{additional metasoft args}
}
\value{
Write the unified p_lrt into the outfile
}
\description{
Combine phenotypes using metasoft
}
