% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pylmm.R
\name{combine_metaSOFT_pylmm}
\alias{combine_metaSOFT_pylmm}
\title{Title}
\usage{
combine_metaSOFT_pylmm(
  infiles,
  midfile,
  outfile,
  version = "2.0.1",
  xargs = ""
)
}
\arguments{
\item{version}{}
}
\value{

}
\description{
Title
}
