% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemma_tools.R
\name{execute_lmm}
\alias{execute_lmm}
\title{Run lmm 2 (LRT) on the data with LOCO}
\usage{
execute_lmm(
  genotypes,
  phenotypes,
  annot,
  covars,
  basedir,
  groups = NULL,
  loco = TRUE,
  runmetasoft = FALSE,
  metasoft_args = ""
)
}
\arguments{
\item{genotypes}{the genotypes data.table}

\item{phenotypes}{the phenotypes data.table}

\item{annot}{annotations of SNPs file}

\item{covars}{covariates data.table}

\item{basedir}{dir to write files to}

\item{groups}{A list of vectors, each containing phenotypes indices to run as multivariate. The return file will be the first group unless runmetasoft is TRUE}

\item{loco}{perform LOCO (default FALSE)}

\item{runmetasoft}{Run metasoft. Return the output file.}

\item{metasoft_args}{additional metasoft arguments}

\item{exec}{gemma executable}
}
\value{
The unified output file
}
\description{
Run lmm 2 (LRT) on the data with LOCO
}
