% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemma_tools.R
\name{get_gemma}
\alias{get_gemma}
\title{Script for executing GEMMA and reading its output
Download gemma if not available, return the executable}
\usage{
get_gemma(basedir, version = "0.98.1")
}
\arguments{
\item{version}{gemma version to download}
}
\value{
the gemma path
}
\description{
Script for executing GEMMA and reading its output
Download gemma if not available, return the executable
}
