% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes.R
\name{get_genes}
\alias{get_genes}
\title{Retrieve genes from biomaRt and return the intersecting genes}
\usage{
get_genes(snps = NULL, dist = 1e+06, attempts = 5, annot = NULL)
}
\arguments{
\item{snps}{A list of SNPs with chr, ps columns. if null returns genes only}

\item{dist}{Distance of genes from SNPs}

\item{attempts}{Maximal number of attempts to try biomaRt}
}
\value{

}
\description{
Retrieve genes from biomaRt and return the intersecting genes
}
