% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MultiTrans.R
\name{get_multi}
\alias{get_multi}
\title{Download MultiTrans}
\usage{
get_multi(
  multilink = "http://genetics.cs.ucla.edu/multiTrans/MultiTrans.zip",
  slidelinux = "http://slide.cs.ucla.edu/data/slide.1.0.4.intel64.tar.gz",
  slidemac = "http://slide.cs.ucla.edu/data/slide.1.0.4.MAC.64bit.tar.gz"
)
}
\arguments{
\item{weblink}{Where to download the zip file from}
}
\value{
TRUE/FALSE if download was successful
}
\description{
Download MultiTrans
}
