% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemma_tools.R
\name{get_residuals}
\alias{get_residuals}
\title{Get resudials of phenotypes
Not used any more}
\usage{
get_residuals(covars, phenotypes)
}
\arguments{
\item{covars}{covariates matrix}

\item{phenotypes}{a table of phenotypes}
}
\value{
a table with residuals of phenotypes
}
\description{
Get resudials of phenotypes
Not used any more
}
