% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemma_tools.R
\name{get_sigmas}
\alias{get_sigmas}
\title{Extract Vg and Ve from GEMMA log file}
\usage{
get_sigmas(logfile)
}
\arguments{
\item{logfile}{GEMMA log file}
}
\value{
Vg and Ve
}
\description{
Extract Vg and Ve from GEMMA log file
}
