% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MultiTrans.R
\name{get_sigmas_dep}
\alias{get_sigmas_dep}
\title{Extract Vg and Ve from GEMMA log file}
\usage{
get_sigmas_dep(logfile)
}
\arguments{
\item{logfile}{GEMMA log file}
}
\value{
Vg and Ve
}
\description{
Extract Vg and Ve from GEMMA log file
}
