% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MultiTrans.R
\name{run_C}
\alias{run_C}
\title{Run java generateC}
\usage{
run_C(basedir, windowSize = 1000)
}
\arguments{
\item{basedir}{where to reasd and write}

\item{windowSize}{default to 1000}
}
\value{

}
\description{
Run java generateC
}
