% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_emma.R
\name{run_emma}
\alias{run_emma}
\title{Run EMMA using the EMMA R package}
\usage{
run_emma(genotypes, phenotypes, annot, covars, basedir, eigens, loco = TRUE)
}
\arguments{
\item{genotypes}{A g (# SNPs) by n + 3 matrix with first three columns rs, major, minor}

\item{phenotypes}{A n by p table (individuals by phenotypes)}

\item{loco}{}
}
\value{
The unified output file
}
\description{
Run EMMA using the EMMA R package
}
