% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inrich_input.R
\name{run_inrich}
\alias{run_inrich}
\title{Rnu KEGG and GO INRICH for written intervals}
\usage{
run_inrich(basedir, name, exec = "inrich", i = 5, j = 200)
}
\arguments{
\item{basedir}{Where the files are}

\item{name}{The phenotype name}

\item{exec}{INRICH executable file}

\item{i}{minimal group size}

\item{j}{maximal group size}
}
\value{

}
\description{
Rnu KEGG and GO INRICH for written intervals
}
