% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pylmm.R
\name{run_pylmm}
\alias{run_pylmm}
\title{Run pylmm}
\usage{
run_pylmm(
  genotypes,
  phenotypes,
  annot,
  covars,
  basedir,
  pylmm,
  pylmm_kinship,
  loco = TRUE,
  metasoft_args = ""
)
}
\arguments{
\item{genotypes}{All genotypes as a data.table, including three columns of rs, major, minor}

\item{phenotypes}{All phenotypes (n x p)}

\item{annot}{SNPs annotations}

\item{covars}{Covariates or null}

\item{basedir}{wd}

\item{pylmm_kinship}{pyLMMKinship executable}

\item{loco}{Use LOCO or not}

\item{metasoft_args}{additional arguments for metasoft}

\item{pymm}{PyLMM executable}
}
\value{
the output file
}
\description{
Run pylmm
}
