% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pylmm.R
\name{run_pylmm_exec}
\alias{run_pylmm_exec}
\title{Run pyLMM for each phenotype and combine using metaSOFT}
\usage{
run_pylmm_exec(
  pylmm,
  geno_sfile,
  annot,
  phenofile,
  ksfile,
  nphen,
  output_head,
  metasoft_args = ""
)
}
\arguments{
\item{pylmm}{pylmmGWAS.py executable}

\item{geno_sfile}{the genotypes file}

\item{annot}{SNP names in the same order in geno_sfile. Names will be replaced in the output file}

\item{phenofile}{the phenotypes file}

\item{ksfile}{kinship matrix file}

\item{nphen}{number of phenotypes}

\item{output_head}{Name of output files prefix}
}
\value{
The combined output file
}
\description{
Run pyLMM for each phenotype and combine using metaSOFT
}
