% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MultiTrans.R
\name{run_R}
\alias{run_R}
\title{Wrap generateR.R}
\usage{
run_R(genotypes, kinship_file, vg_med, ve_med, basedir)
}
\arguments{
\item{kinship_file}{The kinship matrix}

\item{vg_med}{median Vg value}

\item{ve_med}{median Ve value}

\item{basedir}{dir to write r.txt to}

\item{genotypes_file}{The genotypes matrix (just the values)}
}
\value{

}
\description{
Wrap generateR.R
}
