% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MultiTrans.R
\name{run_slide}
\alias{run_slide}
\title{Run SLIDE}
\usage{
run_slide(
  basedir,
  nsnp,
  pval = 0.05,
  windowSize = 1000,
  samps = 1e+05,
  seed = 100
)
}
\arguments{
\item{basedir}{Where to read and write the files}

\item{nsnp}{number of SNPs}

\item{pval}{desired pvalue (default 0.05)}

\item{windowSize}{same as in run_C}

\item{samps}{number of random sampling}

\item{seed}{rand seed}
}
\value{
output thresholds file
}
\description{
Run SLIDE
}
