% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inrich_input.R
\name{write_genes_map}
\alias{write_genes_map}
\title{Write the genes, KEGG and GO mapping for INRICH execution}
\usage{
write_genes_map(basedir)
}
\arguments{
\item{basedir}{Where to write the files to}
}
\value{

}
\description{
Write the genes, KEGG and GO mapping for INRICH execution
}
