% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inrich_input.R
\name{write_inrich_phenotype}
\alias{write_inrich_phenotype}
\title{Write the intervals for INRICH execution}
\usage{
write_inrich_phenotype(snps, basedir, name)
}
\arguments{
\item{snps}{a data.frame or tibble with the columns chr, minps an maxps describinf intervals}

\item{basedir}{Where to write the files to}

\item{name}{Name of the phenotype}
}
\value{

}
\description{
Write the intervals for INRICH execution
}
