% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_inrich_input.R
\name{write_inrich_snps}
\alias{write_inrich_snps}
\title{Write the SNPs map for INRICH execution}
\usage{
write_inrich_snps(snps, basedir)
}
\arguments{
\item{snps}{a data.frame or tibble with the columns chr, minps an maxps describinf intervals}

\item{basedir}{Where to write the files to}
}
\value{

}
\description{
Write the SNPs map for INRICH execution
}
